% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{BackOff}
\alias{BackOff}
\title{Base constructor for BackOff classes.}
\usage{
BackOff(max_timeout = 0)
}
\arguments{
\item{max_timeout}{The maximum amount of time the system can backoff (in seconds)}
}
\description{
Builds a generic BackOff object (can't be used directly.)
}
\examples{
backoff_obj <- BackOff(max_timeout = 5)
}

