% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{calculate_back_off.backoff.base}
\alias{calculate_back_off.backoff.base}
\title{Returns the new back off value.}
\usage{
\method{calculate_back_off}{backoff.base}(x)
}
\description{
This is the key function you want to overload if you wish to create your own
BackOff type.
}

