% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{calculate_back_off.backoff.exponential}
\alias{calculate_back_off.backoff.exponential}
\title{Returns the new back off value.}
\usage{
\method{calculate_back_off}{backoff.exponential}(x)
}
\description{
Returns the new back off value.
}

