% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{delay.backoff.base}
\alias{delay.backoff.base}
\title{Function to retrieve delay}
\usage{
\method{delay}{backoff.base}(x)
}
\description{
Delay will return the following
   > 0, number of seconds until the delay is over
   0 delay is up.  Meaning that you should do your next attempt.
}
\examples{
delay(backoff_obj) #5
}

