% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{Exponential}
\alias{Exponential}
\title{Constructor for Exponential backoff classes.}
\usage{
Exponential(exponent = 1, max_timeout = 0)
}
\arguments{
\item{exponent}{The exponent of the equiation}

\item{max_timeout}{The maximum amount of time the system can backoff (in seconds)}
}
\description{
Builds an Exponential BackOff object. Usage:
   backoff_obj <- Exponential(exponent = 2 )
   On N'th failure delay would be set to:
     1st failure  :  1^2 = 1
     2nd failure  :  2^2 = 4
     3rd failure  :  3^2 = 9
     4th failure  :  4^2 = 16
}
\examples{
backoff_obj <- Exponential(exponent = 5)
}

