% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{failure.backoff.base}
\alias{failure.backoff.base}
\title{Failure will indicicate to the object to increment the current BackOff time.}
\usage{
\method{failure}{backoff.base}(x)
}
\description{
The calculate_back_off function is called to get the time in seconds to wait.
The time waited is time+calculated_back_off time, however it is capped by
max_timeout.
}
\examples{
failure(backoff_obj)
}

