% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{Linear}
\alias{Linear}
\title{Constructor for Linar backoff classes.}
\usage{
Linear(slope = 1, b = 0, max_timeout = 0)
}
\arguments{
\item{slope}{The slope of the line}

\item{b}{The y intercept (or minimal backoff)}

\item{max_timeout}{The maximum amount of time the system can backoff (in seconds)}
}
\description{
Builds a Linear BackOff object. Usage:
   backoff_obj <- Linear(slope = 5, b = 0 )
   On N'th failure delay would be set to:
     y = slope * x + b;
     1st failure  :  5 * count + b = 5 * 1 + 0 = 5
     2nd failure  :  5 * 2 + 0 = 10
     3rd failure  :  5 * 3 + 0 = 15
     4th failure  :  5 * 4 + 0 = 20
}
\examples{
backoff_obj <- Linear(slope = 5, b=2)
}

