% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{Random}
\alias{Random}
\title{Constructor for Random backoff classes.}
\usage{
Random(min = 1, max = 100, max_timeout = 0)
}
\arguments{
\item{min}{The minimum sleep time in seconds}

\item{max}{The maximum sleep time in seconds}

\item{max_timeout}{The maximum amount of time the system can backoff (in seconds)}
}
\description{
Builds an Random BackOff object. Usage:
   backoff_obj <- Random(min = 5, max = 100 )
     On N'th failure delay would be set to:
     1st failure  :  a random number between 5 and 100 inclusive.
                     (5 is a possible value)
     2nd failure  :  a random number between 5 and 100 inclusive.
     3rd failure  :  a random number between 5 and 100 inclusive.
}
\examples{
backoff_obj <- Random(min = 5, max = 100)
}

