% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{reset.backoff.base}
\alias{reset.backoff.base}
\title{A reset function}
\usage{
\method{reset}{backoff.base}(x)
}
\description{
Simply just resets backoff object back to a state in which no "backing off" exists.
}
\examples{
reset(backoff_obj)
}

